refactor absolute path, fix typo
