+++ components/calendar/index.ts
@@ -1,0 +1,2 @@
+export * from "./calendar";
+export * as S from "./calendar.styled";
