--- components/calender/index.ts
@@ -1,2 +1,0 @@
-export * from "./calender";
-export * as S from "./calender.styled";
